requirements.txt
# requirements.txt
# Core dependencies for the DAX / EURO STOXX 50 research project.
# Add version pins once a stable environment is established.

# --- Data ---
pandas>=2.0
numpy>=1.24
yfinance>=0.2.36         # Yahoo Finance data (^GDAXI, ^STOXX50E)
pyarrow>=14.0            # Parquet read/write for caching

# --- Statistical analysis ---
scipy>=1.11
statsmodels>=0.14

# --- Machine learning ---
scikit-learn>=1.3
xgboost>=2.0             # Optional: enable in models_config.yaml

# --- Feature engineering & visualization ---
matplotlib>=3.7
seaborn>=0.13            # Used in correlation heatmaps (notebook 03)

# --- Configuration ---
pyyaml>=6.0

# --- Notebook environment ---
jupyter>=1.0
jupyterlab>=4.0
ipykernel>=6.0
nbstripout>=0.6          # Strip cell outputs before git commit

# --- Testing ---
pytest>=7.4
pytest-cov>=4.1

# --- Utilities ---
python-dateutil>=2.8
tqdm>=4.66               # Progress bars for long downloads
